CLI Fortnte Daily Shop
First commit showing a project I worked on while learning C#, uses a per-existing Fortnite API library that makes the calls for the info required, Binary Tree implementation is an improved version of mine that a friend helped with. 

May update in the future to change it into either a web hostable app, or gui version via avalonia.